pyproject.toml
[tool.poetry]
name = "booksapp"
version = "0.1.0"
description = ""
authors = ["Elephantatech <vivekm.mistry@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^3.0.2"
Flask-SQLAlchemy = "^3.1.1"
python-json-logger = "^2.0.7"
mysql-connector-python = "^8.3.0"
flask-login = "^0.6.3"
werkzeug = "^3.0.1"
pydantic-settings = "^2.2.1"

[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.1.0"
black = "^24.2.0"
flask-testing = "^0.8.1"
pylint-pytest = "^1.1.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint]
max-line-length = 1024
